Add 100ms debate app with roomkit
