+++ Debate_RoomV2/Frontend/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@100mslive/react-sdk": "^0.10.35",
+    "@100mslive/roomkit": "^0.3.0",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
