+++ Debate_RoomV2/Frontend/src/index.js
@@ -3,13 +3,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import { HMSRoomProvider } from "@100mslive/react-sdk";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <HMSRoomProvider>
-      <App />
-    </HMSRoomProvider>
+    <App />
   </React.StrictMode>
 );
